kafka/producer/requirement.txt
kafka-python==2.0.2
sockets
python-jose==3.3.0
mongoengine==0.27.0
pydantic==1.10.5
python-dotenv==0.21.1